fix(lint): resolve clang-tidy findings from full STM32 analysis
False positives suppressed in .clang-tidy:
- Disable insecureAPI.DeprecatedOrUnsafeBufferHandling: memset_s is not
  available in newlib-nano on STM32/ESP32 targets
- Disable performance-no-int-to-ptr: integer casts for hardware memory-mapped
  addresses (flash, UART base registers) are unavoidable in embedded C
- Add TypedefIgnoredRegexp for INA226_/ST7735_ vendor driver types

Real findings fixed:
- esp32_packet_framing.c: cast uint32_t timeout_ms to int for hal_uart_write
- serv_current_monitor.c: explicit (float) cast for sample_period division
- hal_uart.c: (void) snprintf return value (task name buffer always fits)
- ips_display.c: (void) snprintf return values (fixed-size float format)

+++ STM32/Middleware/Features/esp32_packet_framing.c
@@ -454,8 +454,8 @@ uart_driver_status_t stm32_uart_send_packet(const uint8_t *data, size_t length, 
     tx_buffer[tx_index++] = STM32_PACKET_END_MARKER;
     
     // Send packet
-    int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT, 
-                               tx_buffer, tx_index, timeout_ms);
+    int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT,
+                               tx_buffer, tx_index, (int)timeout_ms);
     
     os_mutex_give(state.tx_mutex);
     
@@ -586,7 +586,7 @@ int stm32_uart_send_raw(const uint8_t *data, size_t length, uint32_t timeout_ms)
         return -1;
     }
     
-    int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT, data, length, timeout_ms);
+    int sent = hal_uart_write((hal_uart_port_t)STM32_UART_PORT, data, length, (int)timeout_ms);
     
     os_mutex_give(state.tx_mutex);
